Remove duplication in the localisation tests
